requirements.txt
asgiref==3.2.7
certifi==2020.4.5.2
chardet==3.0.4
defusedxml==0.6.0
Django==3.0.6
django-allauth==0.42.0
django-appconf==1.0.4
django-ckeditor==5.9.0
django-imagekit==4.0.2
django-js-asset==1.2.2
django-model-utils==4.0.0
django-mptt==0.11.0
django-notifications-hq==1.6.0
django-password-reset==2.0
django-taggit==1.3.0
idna==2.9
jsonfield==3.1.0
Markdown==3.2.2
oauthlib==3.1.0
pilkit==2.0
Pillow==7.1.2
Pygments==2.6.1
python3-openid==3.1.0
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
swapper==1.1.2.post1
urllib3==1.25.9
